kci_build: add kernel build commands
Add commands to kci_build for each kernel build step following the new
functions in kernelci.build, namely:

* build_kernel
* install_kernel
* push_kernel
* publish_kernel

The help messages explain how to use them.  The build_kernel command
can be run without a build config name, and with parameters such as
the defconfig can be overridden manually to make it possible to create
debug variants of KernelCI builds.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>