debian: add buster-igt with meson
Replace stretch-igt which used autotools with buster-igt which uses
meson to build libdrm and igt.  This also keeps and installs all the
files from igt as the size is acceptable and we avoid the risk of
failing test cases due to a partial installation.  It also saves
manual updates to the build script each time there is a change in the
test binaries built as part of igt.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>